requirements.txt
fastapi
uvicorn
sqlmodel
asyncpg
selenium
beautifulsoup4
webdriver-manager
python-dotenv